requirements.txt
inky>=2.1.0
Pillow>=9.0.0
requests>=2.25.0